Refactor count methods to use size_t instead of long
- Updated the return type of countBySpace methods in various repositories (MemoryDataAccessControlRepository, MemoryDataFlowRepository, MemoryRemoteTableRepository, MemoryTaskChainRepository, MemoryTaskRepository, MemoryViewRepository) from long to size_t.
- Changed the return type of countByTenant in MemorySpaceRepository from long to size_t.
- Modified count methods in document repositories (MemoryDocumentRepository, MemoryDocumentVersionRepository) to return size_t instead of long.
- Adjusted count methods in document AI repositories (ClientRepository, DocumentRepository, DocumentTypeRepository, etc.) to use size_t.
- Updated count methods in Hana repositories (AlertRepository, BackupRepository, ConfigurationRepository, etc.) to return size_t instead of long.

+++ ai-core/source/uim/platform/ai_core/application/use_cases/manage/artifacts.d
@@ -104,7 +104,7 @@ class ManageArtifactsUseCase : UIMUseCase {
     return CommandResult(true, id.toString, "");
   }
 
-  long count(ResourceGroupId rgId) {
+  size_t count(ResourceGroupId rgId) {
     return repo.countByResourceGroup(rgId);
   }
 }
