+++ ai-core/source/uim/platform/ai_core/infrastructure/persistence/memory/artifacts.d
@@ -70,7 +70,7 @@ class MemoryArtifactRepository : ArtifactRepository {
     }
   }
 
-  long countByResourceGroup(ResourceGroupId rgId) {
+  size_t countByResourceGroup(ResourceGroupId rgId) {
     if (auto rg = rgId in store)
       return cast(long)(*rg).length;
     return 0;
