+++ ai-core/source/uim/platform/ai_core/infrastructure/persistence/memory/executables.d
@@ -58,7 +58,7 @@ class MemoryExecutableRepository : ExecutableRepository {
     }
   }
 
-  long countByScenario(ScenarioId scenarioId, ResourceGroupId rgId) {
+  size_t countByScenario(ScenarioId scenarioId, ResourceGroupId rgId) {
     if (auto rg = rgId in store)
       return cast(long)(*rg).filter!(e => e.scenarioId == scenarioId).array.length;
     return 0;
