+++ ai-core/source/uim/platform/ai_core/infrastructure/persistence/memory/scenarios.d
@@ -63,7 +63,7 @@ class MemoryScenarioRepository : ScenarioRepository {
     }
   }
 
-  long countByResourceGroup(ResourceGroupId rgId) {
+  size_t countByResourceGroup(ResourceGroupId rgId) {
     if (auto rg = rgId in store)
       return cast(long)(*rg).length;
     return 0;
