+++ credential-store/source/uim/platform/credential_store/application/usecases/get_audit_logs.d
@@ -37,7 +37,7 @@ class GetAuditLogsUseCase : UIMUseCase {
     return repo.findById(id);
   }
 
-  long count(TenantId tenantId) {
+  size_t count(TenantId tenantId) {
     return repo.countByTenant(tenantId);
   }
