+++ credential-store/source/uim/platform/credential_store/application/usecases/manage_namespaces.d
@@ -69,7 +69,7 @@ class ManageNamespacesUseCase : UIMUseCase {
     repo.remove(id);
   }
 
-  long count(TenantId tenantId) {
+  size_t count(TenantId tenantId) {
     return repo.countByTenant(tenantId);
   }
