+++ credential-store/source/uim/platform/credential_store/application/usecases/manage_service_bindings.d
@@ -80,7 +80,7 @@ class ManageServiceBindingsUseCase : UIMUseCase {
     repo.remove(id);
   }
 
-  long count(TenantId tenantId) {
+  size_t count(TenantId tenantId) {
     return repo.countByTenant(tenantId);
   }
