+++ credential-store/source/uim/platform/credential_store/domain/ports/repositories/keyring_version_repository.d
@@ -17,5 +17,5 @@ interface KeyringVersionRepository {
   void deactivateAll(CredentialId keyringId);
   void remove(KeyringVersionId id);
   void removeByKeyring(CredentialId keyringId);
-  long countByKeyring(CredentialId keyringId);
+  size_t countByKeyring(CredentialId keyringId);
 }
