+++ datasphere/source/uim/platform/datasphere/infrastructure/persistence/memory/connection_repo.d
@@ -58,7 +58,7 @@ class MemoryConnectionRepository : ConnectionRepository {
     }
   }
 
-  long countBySpace(SpaceId spaceId) {
+  size_t countBySpace(SpaceId spaceId) {
     if (auto sp = spaceId in store)
       return cast(long)(*sp).length;
     return 0;
