+++ datasphere/source/uim/platform/datasphere/infrastructure/persistence/memory/data_access_control_repo.d
@@ -58,7 +58,7 @@ class MemoryDataAccessControlRepository : DataAccessControlRepository {
     }
   }
 
-  long countBySpace(SpaceId spaceId) {
+  size_t countBySpace(SpaceId spaceId) {
     if (auto sp = spaceId in store)
       return cast(long)(*sp).length;
     return 0;
