+++ document-ai/source/uim/platform/document_ai/infrastructure/persistence/memory/document_repo.d
@@ -70,13 +70,13 @@ class MemoryDocumentRepository : DocumentRepository {
     }
   }
 
-  long countByClient(ClientId clientId) {
+  size_t countByClient(ClientId clientId) {
     if (auto cl = clientId in store)
       return cast(long)(*cl).length;
     return 0;
   }
 
-  long countByStatus(DocumentStatus status, ClientId clientId) {
+  size_t countByStatus(DocumentStatus status, ClientId clientId) {
     if (auto cl = clientId in store)
       return cast(long)(*cl).filter!(d => d.status == status).array.length;
     return 0;
