+++ document-ai/source/uim/platform/document_ai/infrastructure/persistence/memory/document_type_repo.d
@@ -58,7 +58,7 @@ class MemoryDocumentTypeRepository : DocumentTypeRepository {
     }
   }
 
-  long countByClient(ClientId clientId) {
+  size_t countByClient(ClientId clientId) {
     if (auto cl = clientId in store)
       return cast(long)(*cl).length;
     return 0;
