+++ document-ai/source/uim/platform/document_ai/infrastructure/persistence/memory/extraction_result_repo.d
@@ -68,7 +68,7 @@ class MemoryExtractionResultRepository : ExtractionResultRepository {
     }
   }
 
-  long countByClient(ClientId clientId) {
+  size_t countByClient(ClientId clientId) {
     if (auto cl = clientId in store)
       return cast(long)(*cl).length;
     return 0;
