+++ document-ai/source/uim/platform/document_ai/infrastructure/persistence/memory/schema_repo.d
@@ -64,7 +64,7 @@ class MemorySchemaRepository : SchemaRepository {
     }
   }
 
-  long countByClient(ClientId clientId) {
+  size_t countByClient(ClientId clientId) {
     if (auto cl = clientId in store)
       return cast(long)(*cl).length;
     return 0;
